# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API Methodly",
        "description": "Dokumentasi API untuk sistem Backend Methodly.",
        "contact": {
            "email": "support@methodly.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Server API Methodly"
        }
    ],
    "paths": {
...
        "/api/v1/forgot-password": {
            "post": {
                "tags": [
                    "Atur Ulang Kata Sandi"
                ],
                "summary": "Kirim tautan atur ulang kata sandi melalui email",
                "operationId": "8a24c9ad448db48934ec72763d79a359",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Alamat email pengguna",
                                        "type": "string",
                                        "format": "email",
                                        "example": "budi@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tautan atur ulang berhasil dikirim",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tautan atur ulang berhasil dikirim"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Kesalahan validasi data"
                    }
                }
            }
        },
        "/api/v1/reset-password": {
            "post": {
                "tags": [
                    "Atur Ulang Kata Sandi"
                ],
                "summary": "Atur ulang kata sandi pengguna",
                "operationId": "e1fec04a24297c4e70c35533b5621842",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "token": {
                                        "description": "Token dari email",
                                        "type": "string",
                                        "example": "TOKEN_DISINI"
                                    },
                                    "email": {
                                        "description": "Alamat email",
                                        "type": "string",
                                        "format": "email",
                                        "example": "budi@example.com"
                                    },
                                    "password": {
                                        "description": "Kata sandi baru",
                                        "type": "string",
                                        "format": "password",
                                        "example": "rahasia_baru123"
                                    },
                                    "password_confirmation": {
                                        "description": "Konfirmasi kata sandi baru",
                                        "type": "string",
                                        "format": "password",
                                        "example": "rahasia_baru123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kata sandi berhasil diatur ulang",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Kata sandi berhasil diatur ulang"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Kesalahan validasi data"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token Bearer untuk akses API",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}